Move secret store under keyring and add migration tests
