+++ src/openhuman/encryption/README.md
@@ -28,4 +28,4 @@ AES-256-GCM at-rest crypto for AI memory storage and the encrypt/decrypt RPC sur
 
 ## Tests
 
-- This domain has no `*_tests.rs` siblings; the underlying crypto round-trips are exercised by `src/openhuman/security/secrets_tests.rs` and the credentials tests, which both cover encrypt/decrypt happy paths and tampered-ciphertext rejection.
+- This domain has no `*_tests.rs` siblings; the underlying crypto round-trips are exercised by `src/openhuman/keyring/encrypted_store_tests.rs` and the credentials tests, which both cover encrypt/decrypt happy paths and tampered-ciphertext rejection.
