+++ src/openhuman/keyring/mod.rs
@@ -1,11 +1,11 @@
-//! OS-keychain backed secret storage with a dev-mode file backend.
+//! OS-keychain backed secret storage with a test/debug file backend override.
 //!
-//! Wraps the [`keyring`] crate (or a file-based backend in dev) to provide a
+//! Wraps the [`keyring`] crate (or an explicit file backend override) to provide a
 //! namespaced, user-scoped interface to secret storage:
 //! - **macOS**: Keychain (prod)
 //! - **Windows**: Credential Manager (prod)
 //! - **Linux**: Secret Service / libsecret (prod)
-//! - **Any platform — dev**: JSON file at `{workspace}/dev-keychain.json`
+//! - **Tests / explicit override**: JSON file at `{workspace}/dev-keychain.json`
 //!
 //! All keys are scoped under a `user_id` parameter so multiple users can
 //! coexist without collision.  The backend entry key format is:
@@ -16,9 +16,8 @@
 //! The backend is chosen **once** at first use, in this priority order:
 //!
 //! 1. `OPENHUMAN_KEYRING_BACKEND` env var: `"os"` | `"file"` | `"mock"`.
-//! 2. `cfg!(debug_assertions)` → `file`.
-//! 3. `OPENHUMAN_APP_ENV` == `"dev"` or `"staging"` → `file`.
-//! 4. Otherwise → `os` (production).
+//! 2. `cfg!(test)` → `file`.
+//! 3. Otherwise → `os`.
 //!
 //! The selected backend is logged once with `[keyring] backend=<name> ...`.
 //!
@@ -30,16 +29,15 @@
 //! always reports as available.
 
 pub mod backend;
+pub mod encrypted_store;
 pub mod error;
 pub mod ops;
 pub mod store;
 
-#[cfg(test)]
-mod tests;
-
 // ── Public re-exports ─────────────────────────────────────────────────────────
 
 pub use backend::KeyringBackend;
+pub use encrypted_store::SecretStore;
 pub use error::KeyringError;
 pub use ops::{
     delete, get, get_or_create_random, is_available, migrate_from_file, set, MigrationOutcome,
